+++ src/Fable.Transforms/Rust/Replacements.fs
@@ -2368,15 +2368,13 @@ let hashSets (com: ICompiler) (ctx: Context) r t (i: CallInfo) (thisArg: Expr op
     | "GetEnumerator", Some c ->
         let ar = Helper.LibCall(com, "HashSet", "entries", t, [ c ])
         Helper.LibCall(com, "Seq", "Enumerable::ofArray", t, [ ar ], ?loc = r) |> Some
-    | ("IsProperSubsetOf" | "IsProperSupersetOf" | "UnionWith" | "IntersectWith" | "ExceptWith" | "IsSubsetOf" | "IsSupersetOf" as meth),
-      Some c ->
-        let meth = Naming.lowerFirst meth
-        Helper.LibCall(com, "Set", meth, t, c :: args, ?loc = r) |> Some
-    // TODO!!!
-    // | "CopyTo"
-    // | "SetEquals"
-    // | "Overlaps"
-    // | "SymmetricExceptWith"
+    | "CopyTo", Some c ->
+        let meth = i.CompiledName |> toLowerFirstWithArgsCountSuffix (c :: args)
+        makeLibModuleCall com r t i "HashSet" meth thisArg args |> Some
+    | ("UnionWith" | "IntersectWith" | "ExceptWith" | "SymmetricExceptWith" | "Overlaps" | "SetEquals"), Some _
+    | ("IsProperSubsetOf" | "IsProperSupersetOf" | "IsSubsetOf" | "IsSupersetOf"), Some _ ->
+        let meth = Naming.lowerFirst i.CompiledName
+        makeLibModuleCall com r t i "HashSet" meth thisArg args |> Some
     | meth, _ ->
         let meth = Naming.removeGetSetPrefix meth |> Naming.lowerFirst
         makeLibModuleCall com r t i "HashSet" meth thisArg args |> Some